simplified for Settings merge
