+++ Everyday/Modules/Settings/Settings/SettingsPresenter.swift
@@ -25,23 +25,6 @@ extension SettingsPresenter: SettingsModuleInput {
 }
 
 extension SettingsPresenter: SettingsViewOutput {
-    func didTapLogoutButton() {
-        interactor.logout { result in
-            DispatchQueue.main.async {
-                switch result {
-                case .success:
-                    self.router.routeToAuthentication()
-                case .failure:
-                    self.view?.showAlert()
-                }
-            }
-        }
-    }
-    
-    func didLoadView() {
-        let viewModel = SettingsViewModel()
-        view?.configure(with: viewModel)
-    }    
 }
 
 extension SettingsPresenter: SettingsInteractorOutput {
