+++ Everyday/Modules/Settings/Settings/SettingsViewController.swift
@@ -1,0 +1,32 @@
+//
+//  SettingsViewController.swift
+//  Everyday
+//
+//  Created by Михаил on 16.02.2024.
+//  
+//
+
+import UIKit
+
+final class SettingsViewController: UIViewController {
+    private let output: SettingsViewOutput
+    private let logoutButton = UIButton(type: .system)
+
+    init(output: SettingsViewOutput) {
+        self.output = output
+
+        super.init(nibName: nil, bundle: nil)
+    }
+
+    @available(*, unavailable)
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override func viewDidLoad() {
+        super.viewDidLoad()        
+    }
+}
+
+extension SettingsViewController: SettingsViewInput {
+}
